requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-huggingface
google-generativeai
faiss-cpu
sentence-transformers
python-dotenv
